Update img plugin to accept all type
